# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://admin.lsp-mpsdm.com",
            "description": "Production API Server"
        },
        {
            "url": "https://admin-staging.lsp-mpsdm.com",
            "description": "Staging API Server"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development API Server"
        }
    ],
    "paths": {
...
        "/api/password/reset": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Reset Password",
                "operationId": "App\\Http\\Controllers\\Api\\PasswordController::reset",
                "requestBody": {
                    "description": "Request Reset Password",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "account email",
                                        "type": "string",
                                        "example": "johndoe@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/password/change": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Change Password",
                "operationId": "App\\Http\\Controllers\\Api\\PasswordController::change",
                "requestBody": {
                    "description": "Request Change Password",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "account email",
                                        "type": "string",
                                        "example": "johndoe@example.com"
                                    },
                                    "password": {
                                        "description": "new password",
                                        "type": "string",
                                        "example": "test1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "name": "Authorization",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}